cap_names: USE HARD-CODED CAPABILITY LIST
Instead of autogenerating cap_names.h, a hard-coded header is used to
support new capabilities even if libcap is compiled on older builders.

When this list needs to be updated, run:
$ cd libcap/libcap && make generate_cap_names_header

and commit cap_names.h.

Note that this header is portable, values passed to libcap functions are
validated at runtime, so it is safe to commit the list.

Signed-off-by: László Várady <laszlo.varady@balabit.com>